fix bugs

+++ app/Http/Controllers/AboutController.php
@@ -11,13 +11,15 @@ class AboutController extends Controller
     public function update(Request $request){
         $about = About::first();
         $about->content = $request->content;
-        
-        unlink(public_path('assets/img/about/' . $about->image));
-        $about->image = $request->file('image')->hashName();
-
-        $image = Image::make($request->file('image'))->resize(530,450);
+        if($about->image){
+            unlink(public_path('assets/img/about/' . $about->image));
+            $about->image = $request->file('image')->hashName();
     
-        $image->save('assets/img/about/'.$about->image);
+            $image = Image::make($request->file('image'))->resize(530,450);
+        
+            $image->save('assets/img/about/'.$about->image);
+        }
+
         $about->save();
         return redirect()->back()->with('success','Le contenu de la page About a été modifié');
     }
